Separate out PyOBD2 code to its own module
